Fix for tool Munch when compiling with Mingw64 for 64-bit target

+++ src/tools/munch.cxx
@@ -47,6 +47,7 @@
 #include <stdlib.h>
 #include <stdint.h>
 #include <stdio.h>
+#include <stddef.h>
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
@@ -821,9 +822,9 @@ char* mystrsep(char** stringp, const char delim) {
   if (n > 0) {
     char* dp = (char*)memchr(mp, (int)((unsigned char)delim), n);
     if (dp) {
-      int nc;
+      ptrdiff_t nc;
       *stringp = dp + 1;
-      nc = (int)((unsigned long)dp - (unsigned long)mp);
+      nc = dp - mp;
       rv = (char*)malloc(nc + 1);
       if (rv) {
         memcpy(rv, mp, nc);
